requirements.txt
fastapi[standard]
python-dotenv
requests
wandb
papermill
ipykernel
pandas
scikit-learn
torch
joblib
uvicorn
sqlalchemy
psycopg2-binary 
passlib[bcrypt]
python-jose
authlib
alembic
requests